requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.1
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.0.12
Werkzeug==0.11.4
argparse==1.2.1
azure-common==1.1.1
azure-nspkg==1.0.0
azure-storage==0.30.0
flask-jsontools==0.1.1-0
futures==3.0.5
itsdangerous==0.24
psycopg2==2.6.1
python-dateutil==2.5.0
requests==2.9.1
six==1.10.0
wsgiref==0.1.2
